requirements.txt
Flask~=2.2.2
Flask-SQLAlchemy
Flask-Migrate
sqlalchemy~=1.4.39
alembic~=1.12.0
python-dateutil~=2.8.2